fix: fixed first hand ranking implementation (#24)
* /feature --
Added evaluation feature logic plus full Test coverage

* chore: inline docs and test renaming


* refactor: rename HandRanking to HandRank and move to separate file

* refactor: update HandRanking references to HandRank in tests + provided new test

* refactor: merged main into branch and deleted deprecated file existence

* refactor: rename HandRanking to HandRank and update references in HandEvaluator and tests

---------
